Change all damage calls to float

+++ Assets/Scripts/Enemies/Base/Enemy.cs
@@ -100,7 +100,7 @@ public class Enemy : MonoBehaviour, IDamageable, IEnemyMoveable, ITriggerCheckab
 
 
 #region health functions
-    public void Damage(int damage)
+    public void Damage(float damage)
     {
        CurrentHealth -= damage;
